==> source/07_HMM_paramextract.qmd <==
---
title: "10_HMM parameters extraction"
#subtitle: Part III - Tuberculosis Preventive Treatment Management
author: "Angelique Porciani"
date: last-modified
date-format: "[Last Updated on] DD MMMM, YYYY"
title-block-banner:  "#4E598C" 
format:
  html:
    code-fold: true
    html-math-method: katex
    code-tools: true
    self-contained: true
editor: visual
editor_options: 
  chunk_output_type: inline
toc: true
execute:
  warning: false
  message: false
  cache: true
---

# Import package/data/function 

```{r}
# Load package 
library(momentuHMM)
library(dplyr)
library(ggplot2)
library(furrr)
library(future)

# Load Data 
data_complete <- readRDS("./output/data_complete.rds")
corres_ttmt_ID <- data_complete %>% select(fullID2, Traitement, sexe)

tib_all_file_rouge_HMM <- readRDS("./output/tib_all_file_rouge_HMM.rds")
tib_all_file_orange_HMM <- readRDS("./output/tib_all_file_orange_HMM.rds")

ts <- tib_all_file_orange_HMM$HMM_res[[1]]
plotStates(ts)

# Function
fun_extract_step <- function(res.HMM){
  res_l <- tryCatch(res.HMM$CIreal$step$est %>% data.frame(), 
                    error = function(e) NA)
  return(res_l)
  
}

fun_TiS_extract <- function(res.HMM){
  res_l <- tryCatch(timeInStates(res.HMM) %>% data.frame(), 
                    error = function(e) NA)
  return(res_l)
}


```


```{r}
# Apply to all individuals 
tib_all_file_orange_HMM <- tib_all_file_orange_HMM %>% 
  mutate(step_est=future_map(HMM_res, fun_extract_step),
         TiS_est=future_map(HMM_res, fun_TiS_extract))

tib_all_file_rouge_HMM <- tib_all_file_rouge_HMM %>% 
  mutate(step_est=future_map(HMM_res, fun_extract_step),
         TiS_est=future_map(HMM_res, fun_TiS_extract))

```

```{r}
# NA check 

tib_all_file_orange_HMM %>% filter(HMM_res== "NA") %>% summarise(n=n())# 3
tib_all_file_rouge_HMM %>% filter(HMM_res== "NA") %>% summarise(n=n())# 6

```

# Step analysis 
## step data extraction and transformation for orange and red concentration 

```{r}
# Recuperation des variable d'interet uniquement pour step et Time in State 

step_df_orange <- tib_all_file_orange_HMM %>% filter(!is.na(HMM_res)) %>% 
  inner_join(corres_ttmt_ID) %>% 
  select(Arena, Lot2, Rpi, Replicat,ArenaN, Experiment, fullID2, step_est, Traitement, sexe) %>% 
  unnest(step_est) %>% 
  mutate(metrics=rep(c("mean", "sd","zeromass"), length.out=714))

step_df_rouge <- tib_all_file_rouge_HMM %>% filter(!is.na(HMM_res)) %>% 
  inner_join(corres_ttmt_ID) %>% 
  select(Arena, Lot2, Rpi, Replicat,ArenaN, Experiment, fullID2, step_est, Traitement, sexe) %>% 
  unnest(step_est) %>% mutate(metrics=rep(c("mean", "sd","zeromass"), length.out=846))

# transfo matrix for sort step estimate values 
mat.step_or <- step_df_orange %>% dplyr::filter(metrics=="mean") %>% 
  select(c(8:10))%>% as.matrix()

mat.step_rou <- step_df_rouge %>% dplyr::filter(metrics=="mean") %>% 
  select(c(8:10))%>% as.matrix()

# sorting values 
for (i in 1:nrow(mat.step_or)){
  mat.step_or[i,] <- sort(mat.step_or[i,])
}

for (i in 1:nrow(mat.step_rou)){
  mat.step_rou[i,] <- sort(mat.step_rou[i,])
}


step_df_orange_orderedSt <- step_df_orange %>% 
  dplyr::filter(metrics=="mean") %>% 
  select(!c(8:10)) %>% 
  cbind(mat.step_or) %>% 
  #mutate_if(is.character,as.factor) %>% 
  pivot_longer(cols =c(11:13), names_to ="State" )

write.csv(step_df_orange_orderedSt, "../output/step_df_orange_orderedSt.csv")

step_df_rouge_orderedSt <- step_df_rouge %>% 
  dplyr::filter(metrics=="mean") %>% 
  select(!c(8:10)) %>% 
  cbind(mat.step_rou) %>% 
  #mutate_if(is.character,as.factor) %>% 
  pivot_longer(cols =c(11:13), names_to ="State" )
write.csv(step_df_rouge_orderedSt, "../output/step_df_rouge_orderedSt.csv")

```

## Plot step data 
Data were sorted to have increasing order st1<st2<st3 

### Plot Orange concentration 
```{r}
# plot for visual check Orange
step_df_orange_orderedSt %>% filter(value<3) %>% 
  ggplot()+
  geom_boxplot(aes(y=value, 
                   x=Traitement, 
                   colour=Traitement, 
                   group=Traitement))+
  theme(legend.position="none")+
  facet_grid(~State)

step_df_orange_orderedSt %>% filter(value<3) %>% 
  ggplot()+
  geom_line(aes(y=value, 
                x=State, 
                colour=fullID2, 
                group=fullID2))+
  theme(legend.position="none")

```

### Plot red concentration 

```{r}

# plot for visual check Rouge
step_df_rouge_orderedSt %>% filter(value<3) %>% 
  ggplot()+
  geom_boxplot(aes(y=value, 
                   x=Traitement, 
                   colour=Traitement, 
                   group=Traitement))+
  theme(legend.position="none")+
  facet_grid(~State)

step_df_rouge_orderedSt %>% filter(value<3) %>% 
  ggplot()+
  geom_line(aes(y=value, 
                x=State, 
                colour=fullID2, 
                group=fullID2))+
  theme(legend.position="none")

```

## Analysis 
### Orange analysis 
```{r}
# Analyse Orange  
filt_or_step <- step_df_orange_orderedSt %>% filter(value<3)
hist(sqrt(filt_or_step$value))


glm_st1 <- glmmTMB(value~Traitement, data=subset(filt_or_step, State=="state.1"), 
                   family=gaussian)

glm_st1.b <- glmmTMB(sqrt(value)~Traitement+(1|Replicat), data=subset(filt_or_step, State=="state.1"), 
                   family=Gamma)

AIC(glm_st1, glm_st1.b)

res <- simulateResiduals(glm_st1.b)
plot(res)

performance(glm_st1.b)

glm_st2.b <- glmmTMB(sqrt(value)~Traitement+(1|Replicat), data=subset(filt_or_step, State=="state.2"), 
                     family=Gamma)

res <- simulateResiduals(glm_st2.b)
plot(res)
summary(glm_st2.b)

glm_st3.b <- glmmTMB(sqrt(value)~Traitement(1|Replicat), data=subset(filt_or_step, State=="state.3"), 
                     family=Gamma)

res <- simulateResiduals(glm_st3.b)
plot(res)


```

### Rouge analysis 

```{r}


# Same for red 
filt_rou_step <- step_df_rouge_orderedSt %>% filter(value<3)

hist(filt_rou_step$value)
hist(sqrt(filt_rou_step$value))
hist(log(filt_rou_step$value))# bof 

plot(filt_rou_step$value)
subST1 <- subset(filt_rou_step, State=="state.3")
hist(subST1$value)


glm_st1.rou <- glmmTMB(value~Traitement, data=subset(filt_rou_step, State=="state.1"), 
                   family=gaussian)

glm_st2.rou <- glmmTMB(sqrt(value)~Traitement, data=subset(filt_rou_step, State=="state.2"), 
                       family=gaussian)
check_distribution(glm_st2.rou)

glm_st1.rou.2 <- glmmTMB(value~Traitement+sexe+(1|Replicat), data=subset(filt_rou_step, State=="state.1"), 
                       family=gaussian)

glm_st1.rou.3 <- glmmTMB(sqrt(value)~Traitement+sexe+(1|Replicat), data=subset(filt_rou_step, State=="state.1"), 
                         family=gaussian)

glm_st1.rou.b <- glmmTMB(sqrt(value)~Traitement+(1|Replicat), data=subset(filt_rou_step, State=="state.1"), 
                     family=Gamma)

glm_st1.rou.c <- glmmTMB(value~Traitement+sexe+(1|Replicat), data=subset(filt_rou_step, State=="state.1"), 
                         family=ziGamma)
check_model(glm_st1.rou.c)
res <- simulateResiduals(glm_st1.rou.c)
plot(res)


glm_st1.rou.b.2 <- glmmTMB(sqrt(value)~Traitement+sexe+(1|Replicat), data=subset(filt_rou_step, State=="state.1"), 
                         family=Gamma)

AIC(glm_st1.rou, glm_st1.rou.2, glm_st1.rou.3)

res <- simulateResiduals(glm_st1.rou)
plot(res)


res <- simulateResiduals(glm_st1.b.rou)
plot(res)

performance(glm_st1.rou.b)
check_model(glm_st1.rou.b)

check_distribution(glm_st1.rou.b)
check_distribution(glm_st1.rou.2)

glm_st2.b <- glmmTMB(sqrt(value)~Traitement+(1|Replicat), data=subset(filt_or_step, State=="state.2"), 
                     family=Gamma)

res <- simulateResiduals(glm_st2.b)
plot(res)
summary(glm_st2.b)

glm_st3.b <- glmmTMB(sqrt(value)~Traitement(1|Replicat), data=subset(filt_or_step, State=="state.3"), 
                     family=Gamma)

res <- simulateResiduals(glm_st3.b)
plot(res)
```

# TiS analysis 

## TiS data extraction for orange and red 
```{r}
TiS_df_orange <- tib_all_file_orange_HMM %>% filter(!is.na(HMM_res)) %>% 
  inner_join(corres_ttmt_ID) %>% 
  select(Arena, Lot2, Rpi, Replicat,ArenaN, Experiment, fullID2, TiS_est, Traitement, sexe) %>% 
  unnest(TiS_est) 
  #mutate(metrics=rep(c("mean", "sd","zeromass"), length.out=714))

TiS_df_rouge <- tib_all_file_rouge_HMM %>% filter(!is.na(HMM_res)) %>% 
  inner_join(corres_ttmt_ID) %>% 
  select(Arena, Lot2, Rpi, Replicat,ArenaN, Experiment, fullID2, TiS_est, Traitement, sexe) %>% 
  unnest(TiS_est) 
  #mutate(metrics=rep(c("mean", "sd","zeromass"), length.out=846))

mat.TiS_rouge <- TiS_df_rouge %>% select(c(8:10))%>% as.matrix()

x=c(0.5, 0.001, 1.5)
order(x)
```


```{r}
rank_states_by_values <- function(data, 
                                  state_cols,      # noms des colonnes state.1, state.2, etc.
                                  value_cols,      # colonnes numériques à classer
                                  metric_filter = NULL, 
                                  metric_col = NULL) {
  # ---- 1. Filtrage facultatif selon une métrique ----
  if (!is.null(metric_filter) && !is.null(metric_col)) {
    data <- data %>% filter(.data[[metric_col]] == metric_filter)
  }
  
  # ---- 2. Extraction des valeurs et classement ----
  mat_values <- data %>%
    select(all_of(value_cols)) %>%
    as.matrix()
  
  m_order <- t(apply(mat_values, 1, order))  # donne l’ordre des colonnes
  
  # ---- 3. Création du nom des colonnes de position ----
  pos_names <- paste0("Pos.st", seq_along(value_cols), ".l")
  colnames(m_order) <- pos_names
  
  # ---- 4. Fusion avec les données ----
  data_ext <- cbind(data, m_order)
  
  # ---- 5. Transformation : positions -> états ----
  data_long <- data_ext %>%
    mutate(across(
      .cols = all_of(pos_names),
      .fns = ~ case_when(
        .x %in% seq_along(state_cols) ~ 
          dplyr::coalesce(!!!setNames(as.list(state_cols), as.character(seq_along(state_cols)))[as.character(.x)]),
        TRUE ~ NA_character_
      ),
      .names = "{str_c('State_', str_extract(.col, '[0-9]+'), '_n')}"
    )) %>%
    pivot_longer(
      cols = starts_with("State_"),
      names_to = "State_new",
      values_to = "value"
    )
  
  return(data_long)
}

# pour orange 
Tis_orange_df <- rank_states_by_values(
  data = df,
  state_cols = c("state.1", "state.2", "state.3"),
  value_cols = c("v1", "v2", "v3"),
  metric_col = "metrics",
  metric_filter = "mean"
)

print(result)
```


```{r}
mat.step_or <- step_df_orange %>% dplyr::filter(metrics=="mean") %>% 
  select(c(8:10))%>% as.matrix()

mat.step_rou <- step_df_rouge %>% dplyr::filter(metrics=="mean") %>% 
  select(c(8:10))%>% as.matrix()

m_order_step_or <- matrix(data=NA, nrow=nrow(mat.step_or), ncol=3, 
                         dimnames=list(NULL, c("Pos.st1.l", "Pos.st2.l","Pos.st3.l")))

m_order_step_rou <- matrix(data=NA, nrow=nrow(mat.step_rou), ncol=3, 
                         dimnames=list(NULL, c("Pos.st1.l", "Pos.st2.l","Pos.st3.l")))

# loop (maybe an apply could do the job in one row)
for (i in 1:nrow(mat.step_or)){
  m_order_step_or[i,] <- order(mat.step_or[i,])
}

for (i in 1:nrow(mat.step_rou)){
  m_order_step_rou[i,] <- order(mat.step_rou[i,])
}
# transform as dataframe 
# m_order_TiS_df <- as.data.frame(m_order_TiS)

# bind to LAM data 
TiS_df_or <- cbind(TiS_df_orange, m_order_step_or)
TiS_df_rou <- cbind(TiS_df_rouge, m_order_step_rou)


TiS_df_or_2 <- TiS_df_or %>%
  mutate(across(
    .cols = c(Pos.st1.l, Pos.st2.l, Pos.st3.l),
    .fns = ~ case_when(
      .x == 1 ~ state.1,
      .x == 2 ~ state.2,
      .x == 3 ~ state.3,
      TRUE ~ NA),
     .names = "{str_c('State_', str_extract(.col, '[0-9]+'), '_n')}")) %>%
  pivot_longer(
    cols = starts_with("State_"),
    names_to = "State_new",
    values_to = "value")

write.csv(TiS_df_or_2, "../output/TiS_df_or_2.csv")


TiS_df_rou_2 <- TiS_df_rou %>%
  mutate(across(
    .cols = c(Pos.st1.l, Pos.st2.l, Pos.st3.l),
    .fns = ~ case_when(
      .x == 1 ~ state.1,
      .x == 2 ~ state.2,
      .x == 3 ~ state.3,
      TRUE ~ NA),
     .names = "{str_c('State_', str_extract(.col, '[0-9]+'), '_n')}")) %>%
  pivot_longer(
    cols = starts_with("State_"),
    names_to = "State_new",
    values_to = "value")
write.csv(TiS_df_rou_2, "../output/TiS_df_rou_2.csv")



TiS_df_or_2 %>% 
ggplot()+
    geom_boxplot(aes(x=Traitement, y=value, colour = Traitement))+
  geom_jitter(aes(x=Traitement, y=value, colour = Traitement))+
  facet_wrap(~State_new)+theme_light()+
  ylab("Proportion of time orange concentration")


TiS_df_rou_2 %>% 
ggplot()+
  geom_boxplot(aes(x=Traitement, y=value, colour = Traitement))+
  geom_jitter(aes(x=Traitement, y=value, colour = Traitement))+
  facet_wrap(~State_new)+theme_light()+
  ylab("Proportion of time in each state (red concentration)")


```

# Analyse TIS ORANGE
Essai des distributions beta, gaussian, lognormal. 
## Choix du modèle 
```{r}
datamod_st1=subset(TiS_df_or_2,State_new=="State_1_n")
datamod_st2=subset(TiS_df_or_2,State_new=="State_2_n")
datamod_st3=subset(TiS_df_or_2,State_new=="State_3_n")

hist(datamod_st1$value)
hist(datamod_st2$value)# contiens des 0 donc transfo obligatoire 
hist(datamod_st3$value)

summary(datamod_st1$value)
summary(datamod_st2$value)
summary(datamod_st3$value)


mod_Tis_orange_1_beta <- glmmTMB(value~Traitement+sexe+(1|Replicat), data=datamod_st1, family=beta_family(link="logit"), zi=~1) 
res <- simulateResiduals(mod_Tis_orange_1_beta)
plot(res)# Nickel 

# 2 
# contiens des 0 donc transfo obligatoire pour appliquer une beta 

datamod_st2$value_star <- (datamod_st2$value * (nrow(datamod_st2) - 1) + 0.5) / nrow(datamod_st2)

hist(datamod_st2$value_star)# Autour de 1 inflation 

mod_Tis_orange_2_beta <- glmmTMB(value_star~Traitement+sexe+(1|Replicat), data=datamod_st2, family=beta_family(link="logit")) 

res <- simulateResiduals(mod_Tis_orange_2_beta)
plot(res)# Pas ok 
# Tweedie 
mod_Tis_orange_2_tw <- glmmTMB(value~Traitement+sexe+(1|Replicat), data=datamod_st2, 
                                    family=tweedie(link="log")) 

res <- simulateResiduals(mod_Tis_orange_2_tw)
plot(res)# toujours pas ok 

# gaussian après transfo sqrt 
mod_Tis_orange_2_nm <- glmmTMB(sqrt(value)~Traitement+sexe+(1|Replicat), data=datamod_st2, 
                                    family=gaussian) 

res <- simulateResiduals(mod_Tis_orange_2_nm)
plot(res)# pas ok 

# Comparaison des modèles sur la base de l'AIC 
AIC(mod_Tis_orange_2_beta,mod_Tis_orange_2_tw,mod_Tis_orange_2_nm)# On garde beta (+petite valeur)

# CCL on garde la beta avec transfo de Smithon and co 
datamod_st3$value_star <- (datamod_st3$value * (nrow(datamod_st3) - 1) + 0.5) / nrow(datamod_st3)
hist(datamod_st3$value_star)

mod_Tis_orange_3_beta <- glmmTMB(value_star~Traitement+sexe+(1|Replicat), data=datamod_st3, family=beta_family(link="logit"), zi=~Traitement) 

res <- simulateResiduals(mod_Tis_orange_3_beta)
plot(res)

check_model(mod_Tis_orange_3_beta)
check_distribution(mod_Tis_orange_3_beta)
```


## Model run 

### TIS State 1
```{r}
# St1 
mod_Tis_orange_1_beta <- glmmTMB(value~Traitement+sexe+(1|Replicat), data=datamod_st1, family=beta_family(link="logit"), zi=~1) 
res <- simulateResiduals(mod_Tis_orange_1_beta)
plot(res)# Nickel

summary(mod_Tis_orange_1_beta)
```

### TIS State 2

```{r}
# St2
mod_Tis_orange_2_beta <- glmmTMB(value_star~Traitement+sexe+(1|Replicat), data=datamod_st2, family=beta_family(link="logit")) 

res <- simulateResiduals(mod_Tis_orange_2_beta)
plot(res)# Pas ok 
summary(mod_Tis_orange_2_beta)
```

### TIS State 3

```{r}
#St3 

mod_Tis_orange_3_beta <- glmmTMB(value_star~Traitement+sexe+(1|Replicat), data=datamod_st3, family=beta_family(link="logit"), zi=~Traitement) 

res <- simulateResiduals(mod_Tis_orange_3_beta)
plot(res)
summary(mod_Tis_orange_3_beta)
```

# Analyse TIS ROUGE 

## Data transfo plot 
```{r}

datamod_st1_red=subset(TiS_df_rou_2,State_new=="State_1_n")
datamod_st2_red=subset(TiS_df_rou_2,State_new=="State_2_n")
datamod_st3_red=subset(TiS_df_rou_2,State_new=="State_3_n")

hist(datamod_st1_red$value)
hist(datamod_st2_red$value)
hist(datamod_st3_red$value)

# contiens des 0-1 donc transfo obligatoire 
summary(datamod_st1_red$value)
summary(datamod_st2_red$value)
summary(datamod_st3_red$value)

# transfo de Smithon and co 
datamod_st1_red$value_star <- (datamod_st1_red$value * (nrow(datamod_st1_red) - 1) + 0.5) / nrow(datamod_st1_red)
datamod_st2_red$value_star <- (datamod_st2_red$value * (nrow(datamod_st2_red) - 1) + 0.5) / nrow(datamod_st2_red)
datamod_st3_red$value_star <- (datamod_st3_red$value * (nrow(datamod_st3_red) - 1) + 0.5) / nrow(datamod_st3_red)

```

## Model run
### TIS State 1

```{r}
mod_Tis_rouge_1_beta <- glmmTMB(value_star~Traitement+sexe+(1|Replicat), data=datamod_st1_red, family=beta_family(link="logit"), zi=~Traitement) 

res <- simulateResiduals(mod_Tis_rouge_1_beta)
plot(res)

check_distribution(mod_Tis_rouge_1_beta)
check_model(mod_Tis_rouge_1_beta)
```

Essai avec Beta reg 
Modelisation des moyennes et des variances separement 
```{r}
library(betareg)
mod_beta1 <- betareg(value_star ~ Traitement +sexe, data = datamod_st1_red)
summary(mod_beta1)

mod_beta2 <- betareg(value_star ~ Traitement +sexe | Traitement + sexe, data = datamod_st1_red)
summary(mod_beta2)
AIC(mod_beta1,mod_beta2)# beta 2 plus petit, on le garde. 

# verif des residus et homoscedasticité. 

resid_std <- residuals(mod_beta2, type = "pearson")
fitted_vals <- fitted(mod_beta2)

plot(fitted_vals, resid_std,
     pch = 19, col = "darkblue",
     xlab = "Valeurs ajustées", ylab = "Résidus de Pearson",
     main = "Résidus vs Valeurs ajustées")
abline(h = 0, col = "red", lty = 2)


library(car)

# Extraire les résidus standardisés
resid_std <- residuals(mod_beta2, type = "pearson")

# Créer un data.frame propre pour le test
df_test <- datamod_st1_red %>%
  drop_na(Traitement, sexe) %>% # enlève les NA
    mutate(resid_std = resid_std) 

# Test de levene 
leveneTest(resid_std ~ interaction(Traitement, sexe), data = df_test)# limite significativité on valide !

ggplot(df_test, aes(x = sexe, y = resid_std, fill = Traitement)) +
  geom_boxplot(alpha = 0.7, outlier.color = "red") +
  theme_minimal() +
  labs(
    title = "Résidus de Pearson par groupe et traitement",
    x = "Groupe",
    y = "Résidus standardisés",
    fill = "Traitement"
  ) +
  theme(plot.title = element_text(face = "bold"))
```
```{r}

mod_beta2_st2 <- betareg(value_star ~ Traitement +sexe | Traitement + sexe, data = datamod_st2_red)

# valid modele 

resid_std <- residuals(mod_beta2, type = "pearson")
fitted_vals <- fitted(mod_beta2)

plot(fitted_vals, resid_std,
     pch = 19, col = "darkblue",
     xlab = "Valeurs ajustées", ylab = "Résidus de Pearson",
     main = "Résidus vs Valeurs ajustées")
abline(h = 0, col = "red", lty = 2)


library(car)

# Extraire les résidus standardisés
resid_std <- residuals(mod_beta2_st2, type = "pearson")

# Créer un data.frame propre pour le test
df_test <- datamod_st2_red %>%
  drop_na(Traitement, sexe) %>% # enlève les NA
    mutate(resid_std = resid_std) 

# Test de levene 
leveneTest(resid_std ~ interaction(Traitement, sexe), data = df_test)# test signif 

ggplot(df_test, aes(x = sexe, y = resid_std, fill = Traitement)) +
  geom_boxplot(alpha = 0.7, outlier.color = "red") +
  theme_minimal() +
  labs(
    title = "Résidus de Pearson par groupe et traitement",
    x = "Groupe",
    y = "Résidus standardisés",
    fill = "Traitement"
  ) +
  theme(plot.title = element_text(face = "bold"))

summary(mod_beta2_st2)
emmeans(mod_beta2_st2, pairwise~Traitement*sexe, infer=TRUE, type="response")
```
